global update

+++ src/pages/photo.tsx
@@ -1,12 +1,14 @@
 import { FC, useState, useEffect } from 'react'
-import { useParams } from 'react-router-dom'
+import { Navigate, useParams } from 'react-router-dom'
 import { usePhoto } from '../hooks/usePhoto'
 import PhotoCard from '../components/photoCard/photoCard'
 import { useAppDispatch, useAppSelector } from '../store/hooks'
-import { deletePhotoFetch, likePhotoFetch, unLikePhotoFetch } from '../store/actions/photosAction'
+import { addCommentFetch, deletePhotoFetch, likePhotoFetch, unLikePhotoFetch } from '../store/actions/photosAction'
 import { addLikePhotoProfile, unLikePhotoProfile } from '../store/actions/profileAction'
 import { ILikesPhoto } from '../types/profileType'
 import { setIsFunAlertIsBoolean } from '../store/slices/pageSlice'
+import { generatorId } from '../helpers/generatorID'
+import { IComment } from '../types/photosType'
 
 const Photo: FC = () => {
   const { id } = useParams()
@@ -14,6 +16,8 @@ const Photo: FC = () => {
   const dispatch = useAppDispatch()
   const profile = useAppSelector(state => state.profileSlice.profile)
   const [isLiked, setIsLiked] = useState<boolean>(false)
+  const [isDelete, setIsDelete] = useState<boolean>(false)
+  const [commentInput, setCommentInput] = useState<string>('')
 
   useEffect(() => {
     if (photo && profile) {
@@ -21,10 +25,46 @@ const Photo: FC = () => {
     }
   })
 
+  const deleteComment = (idComment: string) => {
+    const copyComments = photo?.comments
+
+    console.log(copyComments?.indexOf(copyComments?.find((comment: IComment) => comment.idComment == idComment) ?? {
+      idComment: '',
+      idPhoto: 1,
+      autor: {
+        img: '',
+        nickname: '',
+        id: ''
+      },
+      content: ''}))
+  }
+
+  if (isDelete) {
+    return <Navigate to={'/'} />
+  }
+
+
   return <div>
-    {profile !== null ? <PhotoCard
+    {profile !== null && photo ? <PhotoCard
+      profileAvatar={profile.img}
+      commentInput={commentInput}
+      setCommentInput={setCommentInput}
+      addComment={() => {
+        dispatch(addCommentFetch(photo?.id, [...photo?.comments, {
+          idComment: `${generatorId()}`,
+          idPhoto: photo?.id,
+          autor: {
+            img: profile.img,
+            nickname: profile.nickname,
+            id: profile.id
+          },
+          content: commentInput
+        }]))
+      }}
+      comments={photo?.comments ?? []}
       deletePhoto={() => {
-        dispatch(deletePhotoFetch(id ? +id : 0, profile.id, profile.photos, {id: photo?.id ?? 0, img: photo?.img ?? ''}))
+        dispatch(deletePhotoFetch(id ? +id : 0, profile.id, profile.photos, { id: photo?.id ?? 0, img: photo?.img ?? '' }))
+        setIsDelete(true)
       }}
       idUser={profile.id}
       addLike={() => {
